#2327 adding new files for team types
