+++ src/frontend/src/hooks/teams.hooks.ts
@@ -14,7 +14,6 @@ import {
   deleteTeam,
   createTeam,
   setTeamLeads,
-  setTeamType
 } from '../apis/teams.api';
 
 export interface CreateTeamPayload {
@@ -44,22 +43,6 @@ export const useSetTeamMembers = (teamId: string) => {
     ['teams', 'edit'],
     async (userIds: number[]) => {
       const { data } = await setTeamMembers(teamId, userIds);
-      return data;
-    },
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries(['teams']);
-      }
-    }
-  );
-};
-
-export const useSetTeamType = (teamId: string) => {
-  const queryClient = useQueryClient();
-  return useMutation<{ message: string }, Error, string>(
-    ['teams', 'edit'],
-    async (teamTypeId: string) => {
-      const { data } = await setTeamType(teamId, teamTypeId);
       return data;
     },
     {
